Convert indent action-creator to reducer.
