+++ src/action-creators/index.js
@@ -16,7 +16,6 @@ export { default as expandContextThought } from './expandContextThought'
 export { default as flushQueue } from './flushQueue'
 export { default as home } from './home'
 export { default as importText } from './importText'
-export { default as indent } from './indent'
 export { default as loadFromUrl } from './loadFromUrl'
 export { default as loadLocalThoughts } from './loadLocalThoughts'
 export { default as loadRemoteState } from './loadRemoteState'
