+++ src/shortcuts/indent.js
@@ -5,9 +5,6 @@ import {
   isDocumentEditable,
 } from '../util'
 
-// action-creators
-import indent from '../action-creators/indent'
-
 // eslint-disable-next-line jsdoc/require-jsdoc
 const Icon = ({ fill = 'black', size = 20, style }) => <svg version="1.1" className="icon" xmlns="http://www.w3.org/2000/svg" width={size} height={size} fill={fill} style={style} viewBox="0 0 64 64" enableBackground="new 0 0 64 64">
   <path d="m10 12h44c1.104 0 2-.896 2-2s-.896-2-2-2h-44c-1.104 0-2 .896-2 2s.896 2 2 2z" />
@@ -24,5 +21,5 @@ export default {
   description: `Move the current thought to the end of the previous thought.`,
   svg: Icon,
   canExecute: getState => isDocumentEditable() && getState().cursor,
-  exec: dispatch => dispatch(indent())
+  exec: dispatch => dispatch({ type: 'indent' })
 }
